Persistence in demo mode

+++ main.go
@@ -29,10 +29,13 @@ const LINKSMART = `
 `
 
 var (
-	confPath    = flag.String("conf", "conf/historical-datastore.json", "Historical Datastore configuration file path")
-	profile     = flag.Bool("profile", false, "Enable the HTTP server for runtime profiling")
-	version     = flag.Bool("version", false, "Show the Historical Datastore API version")
-	demomode    = flag.Bool("demo", false, "Run Historical Datasource in demo mode. This creates the service with a growing dummy data")
+	confPath = flag.String("conf", "conf/historical-datastore.json", "Historical Datastore configuration file path")
+	profile  = flag.Bool("profile", false, "Enable the HTTP server for runtime profiling")
+	version  = flag.Bool("version", false, "Show the Historical Datastore API version")
+	demomode = flag.Bool("demo", false, "Run Historical Datastore in demo mode. This creates the service with a growing dummy data.\n"+
+		"By default the data will not be persistent. Inorder to run hds in persistent mode use \"-persistent\" flag")
+	PersistentDemo = flag.Bool("persistent", false, "While running Historical Datastore in demo mode, use persistent storage location specified"+
+		" in the config file")
 	Version     string // set with build flags
 	BuildNumber string // set with build flags
 )
@@ -67,14 +70,21 @@ func main() {
 	}
 
 	if *demomode {
-		conf.Data.Backend.DSN = os.TempDir() + string(os.PathSeparator) + "hds_demo_" + strconv.FormatInt(time.Now().UnixNano(), 10)
+
 		log.Println("===========================")
 		log.Printf("RUNNING IN DEMO MODE")
 		log.Println("===========================")
-		log.Printf("Storing demo data in %s", conf.Data.Backend.DSN)
-		defer os.Remove(conf.Data.Backend.DSN) //remove the temporary file if created on exit
-		//use memory in demo mode for registry
-		conf.Reg.Backend.Type = registry.MEMORY
+
+		if !*PersistentDemo {
+			conf.Data.Backend.DSN = os.TempDir() + string(os.PathSeparator) + "hds_demo_" + strconv.FormatInt(time.Now().UnixNano(), 10)
+			//use memory in demo mode for registry
+			conf.Reg.Backend.Type = registry.MEMORY
+			defer os.Remove(conf.Data.Backend.DSN) //remove the temporary file if created on exit
+		} else {
+			log.Printf("Storing registry data in %s.", conf.Reg.Backend.DSN)
+		}
+		log.Printf("Storing senml data in %s.", conf.Data.Backend.DSN)
+
 	}
 	if conf.ServiceID == "" {
 		conf.ServiceID = uuid.NewV4().String()
